logging-service/requirements.txt
python-logging-loki
